extract interface from DataSourceConstraint
