+++ fili-sql/src/test/groovy/com/yahoo/bard/webservice/sql/builders/SimpleDruidQueryBuilder.groovy
@@ -57,7 +57,7 @@ import com.yahoo.bard.webservice.table.ConstrainedTable
 import com.yahoo.bard.webservice.table.PhysicalTableDictionary
 import com.yahoo.bard.webservice.table.SqlPhysicalTable
 import com.yahoo.bard.webservice.table.availability.PermissiveAvailability
-import com.yahoo.bard.webservice.table.resolver.DataSourceConstraint
+import com.yahoo.bard.webservice.table.resolver.BaseDataSourceConstraint
 import com.yahoo.bard.webservice.util.Utils
 import com.yahoo.bard.webservice.web.filters.ApiFilters
 
@@ -212,7 +212,7 @@ class SimpleDruidQueryBuilder {
         return new TableDataSource(
                 new ConstrainedTable(
                         strictPhysicalTable,
-                        new DataSourceConstraint(
+                        new BaseDataSourceConstraint(
                                 [] as Set,
                                 [] as Set,
                                 [] as Set,
